Some changes in entities, FavoriteListing abstracted.
Buyers and sellers are now able to print themselves in a readable way. Also added a way for BuyerContainer to return information about a specific buyer, and for ListingContainer to return information about a specific listing. Also, modified the type parameters of the containers to make sure we are have the right types in them.