Some refactoring and a new feature
Moved the inbox and outbox attributes to the User abstract class, which means admins now also have inboxes and outboxes. Added a way for buyers and sellers to be able to report each other, and the reports are supposed to go to the inbox of all admins (there is a bug to be fixed about this).