New listing viewing feature and some changes
Added a way for admins to view the listings of a seller, and a new exception to account for the case of the input not being a seller. Additionally moved the logic of getting the listings of a seller out of CreateListing to a new class (ViewListings). Added another test file, ViewListingsTest, to test ViewListings.  Moved the tests that concerned getting the listings of a seller out of CreateListingTest and into ViewListingTest.
Lastly, began the refactoring of return types to be as general as possible. (For example List<String> instead of ArrayList<String>) .