A change in how messages are sent
Moved the process of sending messages from the User class to the Message and ReportMessage classes. The User class now just has the process of adding to the inboxes and outboxes.
Report messages no longer show up in the reporting user's outbox, as their sending process is now different.

This additionally makes sending messages easier, since the method to send the messages takes no parameters.