Added a TODO file, and another feature
Added a way for admins to be able to modify the status of reports, to either resolved or unresolved.
Added a new ViewMessages class that will serve the same purpose as ViewListings, but for messages instead.

Added a facility to obtain a single admin to the UserContainer, and the facility to obtain the reports (only) of an admin in both raw form and string form have been added into ViewMessages.

Changed the way report messages are written to and parsed from the database, to include whether they have been resolved or not.