Bug fixes, started a new feature
FIxed bugs on logging in and on the welcome screen that would cause the program to crash when some invalid inputs were entered

Began implementing the feature that will allow sellers to modify all aspects of their listings (except for the listing ID). Temporarily removed the feature of updating prices.